hell yeah
I am feeling decent about the refactoring and classes and renaming and general cleanliness of the code. Let me know your thoughts.
I want to still implement manual traversal and making the view pane scale in size to the input argument.
Once I have the manual traversal down. I think it would be cool to "hide" the treasure and allow a user a finite number of "digs" to try and find the treasure which could be passed in as args.
Then the question how do we make it a game that is challening but doable but also with some level of objectives.

So like maybe a "warm" "cold" "coldest" "hot" based on the distance from the treasure so its like a metal detector? lmao so its like we give hints but its still a chance a user could lose if they don't dig right and then the difficulty gets harder if you only have like 2 digs and we will only indicate "Warm" which is if the distance is less than 2 blocks away. why the fuck am I still typing right now.