View class added and autosize (sort of)
Separated out GUI components into a View class.
I think conceptually its decent but I may have lost some soundness in other areas in the process so feel free to review. First stab at this so hopefully its not garbage fire.

I got it to scale to input arg for size but my machine is necked at 16x16 so I set the limit on the arg for that for now but we obviously want to look at scaling to the machine's screen size (holy shit lol) and then setting component size properly.