Singleton View, decoupling view and Main
View is now a Singleton, this will help us kind of remove the hard dependency on some of the pieces that were previously passed in (i.e. view, gridPaneSize... etc).  for SetupButtons, we're now passing in Consumer objects, that are basically functions we can invoke.  It makes it cleaner and we won't have to have that dependency on Main