holy fuck the bot is somewhat smart now
added in a move history - so if the move is not one where the bot has already been, it gets added to a priority move queue where it will be made before all "possible" moves. Did testing and it on average reduces the move time for even big boards by a lot! Got a GUI thing to fix on exit game button. And we still need to unblock the GUI on autoplay.

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>