Joel dev (#4)
* push

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>

* holy fuck the bot is somewhat smart now

added in a move history - so if the move is not one where the bot has already been, it gets added to a priority move queue where it will be made before all "possible" moves. Did testing and it on average reduces the move time for even big boards by a lot! Got a GUI thing to fix on exit game button. And we still need to unblock the GUI on autoplay.

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>

* ok GUI issues fixed.

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>

* changes

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>

* does not allow letters on input and optimized the movements a but

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>

* tabbed back view

Co-Authored-By: Matt Smith <queencitysmitty@users.noreply.github.com>

Co-authored-by: Matt Smith <matthew.j.smith2@outlook.com>